src/whisper.local/requirements.txt
fastapi
uvicorn[standard]
faster-whisper
numpy
python-multipart
nvidia-cudnn-cu12==9.1.0.70
pytest
httpx
websockets
pydub
librosa
soundfile
sounddevice
pytest-asyncio
